--- java_app/discord2html/src/main/java/pro/eng/yui/oss/batch/D2HBotRunner.java
@@ -1,5 +1,0 @@
-package pro.eng.yui.oss.batch;
-
-
-public class D2HBotRunner {
-}
