--- java_app/discord2html/src/main/java/pro/eng/yui/oss/botIF/DiscordBot.java
@@ -1,4 +1,0 @@
-package pro.eng.yui.oss.botIF;
-
-public class DiscordBot {
-}
