+++ java_app/discord2html/src/main/java/pro/eng/yui/oss/d2h/DiscordToHtmlBot.java
@@ -1,4 +1,4 @@
-package pro.eng.yui.oss;
+package pro.eng.yui.oss.d2h;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
