+++ java_app/discord2html/src/main/java/pro/eng/yui/oss/d2h/batch/D2HBotRunner.java
@@ -1,0 +1,5 @@
+package pro.eng.yui.oss.d2h.batch;
+
+
+public class D2HBotRunner {
+}
