+++ java_app/discord2html/src/main/java/pro/eng/yui/oss/d2h/botIF/DiscordBot.java
@@ -1,0 +1,4 @@
+package pro.eng.yui.oss.d2h.botIF;
+
+public class DiscordBot {
+}
