+++ java_app/discord2html/src/main/java/pro/eng/yui/oss/d2h/html/FileGenerator.java
@@ -1,0 +1,4 @@
+package pro.eng.yui.oss.d2h.html;
+
+public class FileGenerator {
+}
