--- java_app/discord2html/src/main/java/pro/eng/yui/oss/db/DbConfig.java
@@ -1,4 +1,0 @@
-package pro.eng.yui.oss.db;
-
-public class DbConfig {
-}
