--- java_app/discord2html/src/main/java/pro/eng/yui/oss/github/GitHubManager.java
@@ -1,4 +1,0 @@
-package pro.eng.yui.oss.github;
-
-public class GitHubManager {
-}
