--- java_app/discord2html/src/main/java/pro/eng/yui/oss/html/FileGenerator.java
@@ -1,4 +1,0 @@
-package pro.eng.yui.oss.html;
-
-public class FileGenerator {
-}
